move playlist view to its own component
